Gitignore, create a Base file and move advertiser
